Properly check for existence of files

+++ opensuse/script/update.sh
@@ -7,5 +7,7 @@ if [ "$build_type" = "desktop" ]; then
   zypper --non-interactive install -t pattern xfce
 fi
 
-cd /usr/bin
-ln -s /sbin/shutdown
+if [ ! -f /usr/bin/shutdown ]; then
+  cd /usr/bin
+  ln -s /sbin/shutdown
+fi
